Added comments and better clarification on what transVoxels is doing
