+++ Terrain/Engine/Shaders/Includes/TransvoxelTables.hlsl
@@ -1,4 +1,4 @@
-    
+
 struct RegularCellData
 {
     uint VertexCount;
@@ -34,31 +34,21 @@ StructuredBuffer<uint> RegularCellIndices;
 StructuredBuffer<VertexData> RegularVertexRanges;
 StructuredBuffer<uint> RegularVertexData;
 
-uint GetIndex(RegularCellData cell, uint i)
-{
-    // Caller guarantees: i < IndicesCount
-    return RegularCellIndices[cell.IndicesStart + i];
-}
-
 uint GetPackedVertexU16(uint classId, uint vertexIndex)
 {
-    // vertexIndex = local vertex index (0..VertexCount-1)
     uint baseIndex = RegularVertexRanges[classId].VertexStart;
-    uint packed = RegularVertexData[baseIndex + vertexIndex];
-
-    // Mask to original ushort
-    return packed & 0xFFFF;
+    return RegularVertexData[baseIndex + vertexIndex];
 }
 
 VertexDataUnpacked UnpackVertex(uint packed)
 {
     VertexDataUnpacked v;
-    packed &= 0xFFFF; // safety
+    packed &= 0xFFFF;
 
-    v.Corner1 = packed & 0x0F; // bits 0..3
-    v.Corner0 = (packed >> 4) & 0x0F; // bits 4..7
-    v.CacheIndex = (packed >> 8) & 0x0F; // bits 8..11
-    v.CacheDir = (packed >> 12) & 0x07; // bits 12..15
+    v.Corner1 = packed & 0x0F;
+    v.Corner0 = (packed >> 4) & 0x0F;
+    v.CacheIndex = (packed >> 8) & 0x0F;
+    v.CacheDir = (packed >> 12) & 0x07;
 
     return v;
 }
